Refactor v-form ai usage to composable
